New CI script for versioning bundle of Cargo workspace crates/packages (from Cargo metadata)
